% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_positions.R
\name{all_positions}
\alias{all_positions}
\title{Get all aircraft positions}
\usage{
all_positions()
}
\value{
A data frame of aircraft positions with one row per aircraft. The
row names are aircraft IDs. Returns an empty data frame if there are no
aircraft in the simulation.
}
\description{
Get all aircraft positions
}
\examples{
\dontrun{
all_positions()
}

}
