% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_altitude.R
\name{change_altitude}
\alias{change_altitude}
\title{Send a "change altitude" command}
\usage{
change_altitude(aircraft_id, altitude = NULL, flight_level = NULL,
  vertical_speed = NULL)
}
\arguments{
\item{aircraft_id}{A string aircraft identifier}

\item{altitude}{A double in the range [0, 6000]. The aircraft's altitude in feet.
For altitudes in excess of 6000ft a flight level should be specified instead.}

\item{flight_level}{A integer of 60 or more. The aircraft's flight level.}

\item{vertical_speed}{A non-negative double. The requested vertical speed in feet/min.}
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Request an aircraft to change altitude.
}
\details{
Either the \code{altitude} or \code{flight_level} argument must be given,
but not both.
}
\examples{
\dontrun{
change_altitude("test1234", flight_level = 450)
change_altitude("test1234", altitude = 5000)
}
}
