% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_heading.R
\name{change_heading}
\alias{change_heading}
\title{Send a "change heading" command}
\usage{
change_heading(aircraft_id, heading)
}
\arguments{
\item{aircraft_id}{A string aircraft identifier}

\item{heading}{A double in the range [0, 360). The aircraft's new heading in degrees.}
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Request an aircraft to change heading.
}
\examples{
\dontrun{
change_heading("test1234", heading = 90)
}
}
