% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_speed.R
\name{change_speed}
\alias{change_speed}
\title{Send a "change speed" command}
\usage{
change_speed(aircraft_id, speed)
}
\arguments{
\item{aircraft_id}{A string aircraft identifier}

\item{speed}{A non-negative double. The aircraft's new speed in knots.}
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Request an aircraft to change speed.
}
\examples{
\dontrun{
change_speed("test1234", speed = 90)
}
}
