% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_param.R
\name{config_param}
\alias{config_param}
\title{Get a configuration parameter}
\usage{
config_param(param, file = config_file_location())
}
\arguments{
\item{param}{The name of the config parameter.}

\item{file}{The configuration file to read from.}
}
\value{
The value of the requested configuration parameter. An error is
thrown if the given parameter name is not found in the config file.
}
\description{
Get a configuration parameter
}
