% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_endpoint_url.R
\name{construct_endpoint_url}
\alias{construct_endpoint_url}
\title{Construct a Bluebird API endpoint URL}
\usage{
construct_endpoint_url(endpoint, query = NULL)
}
\arguments{
\item{endpoint}{A string specifying the API endpoint.}

\item{query}{A list of HTTP query parameters.}
}
\value{
A string.
}
\description{
Construct a Bluebird API endpoint URL
}
\examples{
construct_endpoint_url(endpoint = "ic")

}
