% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scenario.R
\name{create_scenario}
\alias{create_scenario}
\title{Create a scenario}
\usage{
create_scenario(filename, scenario)
}
\arguments{
\item{filename}{The path to a scenario file on the local machine.}

\item{scenario}{The name to store scenario under on the simulator host.}
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Upload a scenario file from the local machine to the running simulator.
}
\examples{
\dontrun{
create_scenario(system.file("dodo-test-scenario.scn", package = "rdodo"),
scenario = "dodo-test-scenario")
}

}
