% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_to_waypoint.R
\name{direct_to_waypoint}
\alias{direct_to_waypoint}
\title{Send a "go directly to a waypoint" command}
\usage{
direct_to_waypoint(aircraft_id, waypoint_name)
}
\arguments{
\item{aircraft_id}{A string aircraft identifier.}

\item{waypoint_name}{The name of the waypoint.}
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Request aircraft to change heading toward a given waypoint. The waypoint must
exist on the aircraft route.
}
\examples{
\dontrun{
direct_to_waypoint("TST1000", waypoint_name = "TESTWPT")
}
}
