% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_log.R
\name{episode_log}
\alias{episode_log}
\title{Download the episode log file}
\usage{
episode_log(path = "")
}
\arguments{
\item{path}{(Optional) The directory path under which the log file will be saved.}
}
\value{
The relative path to the saved episode log file, if successful. Otherwise an
error is thrown.
}
\description{
Downloads the episode log and saves it to file in the given \code{path}, or
the working directory by default, under a subdirectory named \code{logs}.
}
\examples{
\dontrun{
episode_log()
}

}
