% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/found_bluebird.R
\name{found_bluebird}
\alias{found_bluebird}
\title{Check communication with BlueBird}
\usage{
found_bluebird()
}
\value{
A boolean, \code{TRUE} indicates a valid response was received from
the BlueBird API.
}
\description{
Check communication with BlueBird
}
\examples{
\dontrun{
found_bluebird()
}

}
