% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_distance.R
\name{geodesic_distance}
\alias{geodesic_distance}
\title{Geodesic distance}
\usage{
geodesic_distance(from_lat, from_lon, to_lat, to_lon)
}
\arguments{
\item{from_lat}{A double in the range [-90, 90].}

\item{from_lon}{A double in the range [-180, 180).}

\item{to_lat}{A double in the range [-90, 90].}

\item{to_lon}{A double in the range [-180, 180).}
}
\value{
A double, the geodesic distance between the two points.
}
\description{
Compute the geodesic distance in metres between two latitude/longitude
positions.
}
