% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bluebird_url.R
\name{get_bluebird_url}
\alias{get_bluebird_url}
\title{Get the URL of the bluebird API}
\usage{
get_bluebird_url()
}
\value{
A string.
}
\description{
Get the URL of the bluebird API
}
\examples{
get_bluebird_url()

}
