% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_call.R
\name{get_call}
\alias{get_call}
\title{Make a GET request to the Bluebird API}
\usage{
get_call(endpoint, query = NULL, validate = TRUE)
}
\arguments{
\item{endpoint}{The Bluebird API endpoing to call.}

\item{query}{A list of HTTP query parameters.}

\item{validate}{A boolean flag. If \code{TRUE} (the default), the HTTP response will be
validated before being returned.}
}
\value{
A \code{\link{response}} object, if successful. Otherwise an error is thrown.
}
\description{
Make a GET request to the Bluebird API
}
\examples{
\dontrun{
endpoint <- config_param("endpoint_simulation_time")
get_call(endpoint = endpoint)
}

}
