% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great_circle_distance.R
\name{great_circle_distance}
\alias{great_circle_distance}
\title{Great circle distance}
\usage{
great_circle_distance(from_lat, from_lon, to_lat, to_lon)
}
\arguments{
\item{from_lat}{A double in the range [-90, 90].}

\item{from_lon}{A double in the range [-180, 180).}

\item{to_lat}{A double in the range [-90, 90].}

\item{to_lon}{A double in the range [-180, 180).}
}
\value{
A double with units, the great circle distance between the two points in metres.
}
\description{
Compute the great circle distance in metres between two latitude/longitude
positions, assuming a radius of the earth of 6378137m.
}
