% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great_circle_separation.R
\name{great_circle_separation}
\alias{great_circle_separation}
\title{Great circle separation}
\usage{
great_circle_separation(from_aircraft_id,
  to_aircraft_id = from_aircraft_id)
}
\arguments{
\item{from_aircraft_id}{A string vector of aircraft IDs.}

\item{to_aircraft_id}{(Optional) A string vector of aircraft IDs. Defaults to
\code{from_aircraft_id}.}
}
\value{
A dataframe, with \code{from_aircraft_id} as row names and
\code{to_aircraft_id} as column names, containing the great circle distance in
metres between the positions of the corresponding pair of aircraft.
}
\description{
Compute the great circle separation in metres between the positions of pairs of
aircraft.
}
