% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause_simulation.R
\name{pause_simulation}
\alias{pause_simulation}
\title{Pause the simulation}
\usage{
pause_simulation()
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Pause the simulation
}
\examples{
\dontrun{
pause_simulation()
}

}
