% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_call.R
\name{position_call}
\alias{position_call}
\title{Make a call to the aircraft position endpoint}
\usage{
position_call(aircraft_id)
}
\arguments{
\item{aircraft_id}{A scalar string aircraft identifier, or the special string \code{all}.}
}
\value{
A list of aircraft position information, or an empty list if the given
\code{aircraft_id} did not match any aircraft in the simulation (this includes
the case that \code{aircraft_id} was the special string \code{all} but there
were no existing aircraft).
}
\description{
Make a call to the aircraft position endpoint
}
\examples{
\dontrun{
position_call('ABC123')
}

}
