% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume_simulation.R
\name{resume_simulation}
\alias{resume_simulation}
\title{Resume the simulation after a pause}
\usage{
resume_simulation()
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Resume the simulation after a pause
}
\examples{
\dontrun{
resume_simulation()
}

}
