% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_call.R
\name{route_call}
\alias{route_call}
\title{Make a call to the aircraft route endpoint}
\usage{
route_call(aircraft_id)
}
\arguments{
\item{aircraft_id}{A scalar aircraft identifier (string).}
}
\value{
A list with three elements: 1. the aircraft ID, 2. route information (a list
of lists) 3. the simulator  time. If the \code{aircraft_id} was not found in the
simulation, the 'route' element will contain a single \code{NA} value. If the
\code{aircraft_id} was found, but the aircraft has no route, the 'route'
element will be an empty list.
}
\description{
Make a call to the aircraft route endpoint
}
\examples{
\dontrun{
route_call('ABC123')
}

}
