% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\name{separation}
\alias{separation}
\title{Separation}
\usage{
separation(from_aircraft_id, to_aircraft_id = from_aircraft_id,
  f_distance)
}
\arguments{
\item{from_aircraft_id}{A string vector of aircraft IDs.}

\item{to_aircraft_id}{(Optional) A string vector of aircraft IDs. Defaults to
\code{from_aircraft_id}.}

\item{f_distance}{A distance function such as\link{geodesic_distance} or
\link{great_circle_distance}.}
}
\value{
A dataframe, with \code{from_aircraft_id} as row names and
\code{to_aircraft_id} as column names, containing the separation between the
positions of the corresponding pair of aircraft.
}
\description{
Compute the separation between the positions of pairs of aircraft under a
given distance measure.
}
