% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_simulation_rate_multiplier.R
\name{set_simulation_rate_multiplier}
\alias{set_simulation_rate_multiplier}
\title{Set the simulation rate multiplier.}
\usage{
set_simulation_rate_multiplier(multiplier)
}
\arguments{
\item{multiplier}{A position double}
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Sets the simulation rate multiplier for the current simulation. By default
this multiplier is equal to one (real-time operation). If set to another
value, the simulation will run faster (or slower) than real-time, with a
fixed multiplier as provided. For example, a multiplier of 2 would cause the
simulation to run twice as fast: 60 simulation minutes take 30 actual minutes.
}
\examples{
\dontrun{
set_simulation_rate_multiplier(2)
set_simulation_rate_multiplier(0.5)
}
}
