% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_time.R
\name{simulation_time}
\alias{simulation_time}
\title{Get the current simulation time}
\usage{
simulation_time()
}
\value{
A \code{POSIXct} date-time object, if successful. Otherwise an error is thrown.
}
\description{
Get the current simulation time
}
\examples{
\dontrun{
simulation_time()
}

}
