% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_sector.R
\name{upload_sector}
\alias{upload_sector}
\title{Upload a sector}
\usage{
upload_sector(filename, sector_name)
}
\arguments{
\item{filename}{The path to a sector GeoJSON file on the local machine.}

\item{sector_name}{The name to store the sector under.}
}
\value{
A boolean, \code{TRUE} indicates that the sector was uploaded
successfully.
}
\description{
Upload a sector
}
\examples{
\dontrun{
filename <- system.file("dodo-test-sector.geojson", package="rdodo")
sector_name <- "test-sector"
upload_sector(filename, sector_name = "test-sector")
}

}
