% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_response.R
\name{validate_response}
\alias{validate_response}
\title{Validate a response from the Bluebird API}
\usage{
validate_response(response)
}
\arguments{
\item{response}{An \code{httr} response object.}
}
\description{
Generates an error in case the response is invalid.
}
\examples{
\dontrun{
response <- httr::POST(construct_endpoint_url(config::get("endpoint_reset_simulation")))
validate_response(response)
}

}
