% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aircraft.R
\name{create_aircraft}
\alias{create_aircraft}
\title{Create an aircraft}
\usage{
create_aircraft(
  aircraft_id,
  type,
  latitude,
  longitude,
  heading,
  altitude = NULL,
  flight_level = NULL,
  speed
)
}
\arguments{
\item{aircraft_id}{A string aircraft identifier}

\item{type}{A string ICAO aircraft type designator}

\item{latitude}{A double in the range [-90, 90]. The aircraft's latitude.}

\item{longitude}{A double in the range [-180, 180). The aircraft's longitude.}

\item{heading}{An integer in the range [0, 360). The aircraft's heading in degrees.}

\item{altitude}{The aircraft's altitude in feet expressed as a double in the range [0, 6000]
or a quantity whose units can be converted to feet. For altitudes in excess
of 6000ft a flight level should be specified instead.}

\item{flight_level}{A integer of 60 or more. The aircraft's flight level.}

\item{speed}{The aircraft's speed in knots (KCAS) expressed as a non-negative double or
a quantity whose units can be converted to knots.}
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Either the \code{altitude} or \code{flight_level} argument must be given,
but not both.
}
\details{
If an aircraft with the given \code{aircraft_id} already exists an error is
thrown.
}
\examples{
\dontrun{
create_aircraft("test1234", "B744", 0, 0, 0, flight_level = 250, speed = 200)
}
}
