% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/found_bluebird.R
\name{found_bluebird}
\alias{found_bluebird}
\title{Check communication with BlueBird}
\usage{
found_bluebird(verbose = FALSE)
}
\arguments{
\item{verbose}{A boolean flag. If \code{TRUE} then an error message is printed to the
console in case Bluebird is not found.}
}
\value{
A boolean, \code{TRUE} indicates that a request to the BlueBird URL
was successful.
}
\description{
Check communication with BlueBird
}
\examples{
\dontrun{
found_bluebird()
}

}
