% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scenario.R
\name{load_scenario}
\alias{load_scenario}
\title{Load a scenario}
\usage{
load_scenario(filename, multiplier = NULL)
}
\arguments{
\item{filename}{The relative path to a scenario file.}

\item{multiplier}{(Optional) A positive number specifying the simulation rate multiplier. If
omitted, the rate multiplier will be 1.}
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Load a scenario
}
\examples{
\dontrun{
load_scenario("scenario/8.SCN")
}

}
