% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scenario.R
\name{load_scenario}
\alias{load_scenario}
\title{Load a scenario}
\usage{
load_scenario(filename)
}
\arguments{
\item{filename}{A string indicating the relative path to a scenario file.}
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Load a scenario
}
\examples{
\dontrun{
load_scenario("scenario/8.SCN")
}

}
