% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_call.R
\name{position_call}
\alias{position_call}
\title{Make a call to the aircraft position endpoint}
\usage{
position_call(aircraft_id = NULL)
}
\arguments{
\item{aircraft_id}{A scalar aircraft identifier (string), or NULL (the default).}
}
\value{
A list of lists, each element named by a \code{aircraft_id} and each sublist
containing aircraft position information, plus an element for the
\code{simulator_time}. If the \code{aircraft_id} was not found in the
simulation, the sublist will be empty. Returns an empty list if
\code{aircraft_id} was NULL but there were no existing aircraft.
}
\description{
Make a call to the aircraft position endpoint
}
\examples{
\dontrun{
position_call('ABC123')
position_call()
}

}
