% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_call.R
\name{position_call}
\alias{position_call}
\title{Make a call to the aircraft position endpoint}
\usage{
position_call(aircraft_id)
}
\arguments{
\item{aircraft_id}{A string aircraft identifier. If this aircraft does not exist an exception
will be thrown.}
}
\value{
A list of aircraft position information.
}
\description{
Make a call to the aircraft position endpoint
}
\examples{
\dontrun{
position_call('ABC123')
}

}
