% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_call.R
\name{post_call}
\alias{post_call}
\title{Make a POST request to the Bluebird API}
\usage{
post_call(endpoint, body = NULL)
}
\arguments{
\item{endpoint}{The Bluebird API endpoing to call.}

\item{body}{A named list.}
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Make a POST request to the Bluebird API
}
\examples{
\dontrun{
body <- list(acid="BA10", type="B744", lat=0, lon=0, hdg=0, alt=20000, spd=240)
endpoint <- config_param("endpoint_create_aircraft")
post_call(endpoint = endpoint, body = body)
}

}
