% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_simulation.R
\name{reset_simulation}
\alias{reset_simulation}
\title{Reset the simulation}
\usage{
reset_simulation()
}
\value{
\code{TRUE} if successful. Otherwise \code{FALSE} and an error is thrown.
}
\description{
Reset the simulation
}
\examples{
\dontrun{
reset_simulation()
}

}
