% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_distance.R
\name{vertical_distance}
\alias{vertical_distance}
\title{Vertical distance}
\usage{
vertical_distance(from_alt, to_alt)
}
\arguments{
\item{from_alt}{A non-negative double, interpreted as an altitude in feet, or a double value
with a distance unit.}

\item{to_alt}{A non-negative double, interpreted as an altitude in feet, or a double value
with a distance unit.}
}
\value{
A double with units, the vertical distance between the two altitudes in metres.
}
\description{
Compute the vertical distance in metres between two altitudes. This is a
trivial helper function for \code{\link{vertical_separation}}.
}
